requirements.txt
requests
trafilatura
feedparser
psycopg2
python-dotenv
